Consolidate index file exports
