--- src/components/buttons/index.js
@@ -1,2 +1,0 @@
-export * from './Button';
-export * from './LinkedButton';
