--- src/components/category/index.js
@@ -1,2 +1,0 @@
-export * from './Categories';
-export * from './CategoryItem';
