--- src/components/home/index.js
@@ -1,2 +1,0 @@
-export * from './About';
-export * from './HeroSection';
