+++ src/components/index.js
@@ -1,0 +1,16 @@
+export * from './buttons/Button';
+export * from './buttons/LinkedButton';
+export * from './category/Categories';
+export * from './category/CategoryItem';
+export * from './home/About';
+export * from './home/HeroSection';
+export * from './layout/Footer';
+export * from './layout/Layout';
+export * from './layout/Navbar';
+export * from './mealsPage/IngredientsTable';
+export * from './mealsPage/SearchBar';
+export * from './mealsPage/SingleMealCard';
+export * from './text/PointerText';
+export * from './text/Text';
+export * from './text/Title';
+
