+++ src/components/layout/Layout.js
@@ -1,4 +1,4 @@
-import { Footer, Navbar } from "./";
+import { Footer, Navbar } from "@/components";
 import React, { Fragment } from "react";
 import classes from './Layout.module.scss';
