--- src/components/layout/index.js
@@ -1,2 +1,0 @@
-export * from './Footer';
-export * from './Navbar';
