--- src/components/mealsPage/index.js
@@ -1,3 +1,0 @@
-export * from './IngredientsTable';
-export * from './SearchBar';
-export * from './SingleMealCard';
