--- src/components/text/index.js
@@ -1,3 +1,0 @@
-export * from './PointerText';
-export * from './Text';
-export * from './Title';
